+++ pyrgg/params.py
@@ -4,8 +4,8 @@ import os
 
 MENU_ITEMS1 = {
     "file_name": "File Name : ",
-    "output_format": "Graph Format : \nDIMACS(.gr)[1] | JSON(.json)[2] | CSV(.csv)[3] | YAML(.yaml)[4]\n| WEL(.wel)[5] | ASP(.lp)[6] | Pickle(.p)[7] | UCINET DL Format(.dl)[8] | TGF(.tgf)[9] | TSV("
-                     ".tsv)[10]",
+    "output_format": "Graph Formats : \n1- DIMACS(.gr)\n2- JSON(.json)\n3- CSV(.csv)\n4- YAML(.yaml)\n5- WEL(.wel)\n6- ASP(.lp)\n7- Pickle(.p)\n8- UCINET DL(.dl)\n9- TGF(.tgf)\n10- TSV("
+                     ".tsv)\n11- Matrix Market(.mtx)\n",
     "weight": "Weighted[1] or Unweighted[2]"}
 
 MENU_ITEMS2 = {"vertices": "Vertices Number : ",
@@ -29,10 +29,11 @@ SUFFIX_MENU = {
     7: ".p",
     8: ".dl",
     9: ".tgf",
-    10: ".tsv"}
+    10: ".tsv",
+    11: ".mtx"}
 
 
-PYRGG_VERSION = "0.5"
+PYRGG_VERSION = "0.6"
 
 SOURCE_DIR = os.getcwd()