+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.6] - 2020-07-24
+### Added
+- Matrix Market format
+### Changed
+- `json_maker` function optimized
+- `dl_maker` function optimized
+- `tgf_maker` function optimized
+- `lp_maker` function optimized
 ## [0.5] - 2020-07-01
 ### Added
 - TSV format
@@ -57,7 +65,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - JSON format
 - README
 
-[Unreleased]: https://github.com/sepandhaghighi/pyrgg/compare/v0.5...dev
+[Unreleased]: https://github.com/sepandhaghighi/pyrgg/compare/v0.6...dev
+[0.6]: https://github.com/sepandhaghighi/pyrgg/compare/v0.5...v0.6
 [0.5]: https://github.com/sepandhaghighi/pyrgg/compare/v0.4...v0.5
 [0.4]: https://github.com/sepandhaghighi/pyrgg/compare/v0.3...v0.4
 [0.3]: https://github.com/sepandhaghighi/pyrgg/compare/v0.2...v0.3